+++ app/controllers/application_controller.rb
@@ -1,5 +1,6 @@
 class ApplicationController < ActionController::Base
   TYPING_ACTIVITY_SESSION_KEY = "typing_last_activity_at"
+  TYPING_LOCK_RETURN_TO_SESSION_KEY = "typing_lock_return_to"
   LEGACY_TYPING_UNLOCK_SESSION_KEY = "typing_unlocked_at"
   TYPING_AUTHENTICATOR_PENDING_SESSION_KEY = "typing_authenticator_pending_at"
   TYPING_AUTHENTICATOR_RETURN_TO_SESSION_KEY = "typing_authenticator_return_to"
@@ -22,7 +23,12 @@ class ApplicationController < ActionController::Base
     target = safe_return_path(request.fullpath)
 
     if @user.typing_fingerprint_configured?
-      redirect_for_typing_lock(typing_lock_path(return_to: target))
+      if typing_lock_root_display_request?
+        render_typing_lock_unlock(return_to: typing_lock_return_to_path(target))
+      else
+        remember_typing_lock_return_to!(target)
+        redirect_for_typing_lock(root_path)
+      end
     else
       redirect_for_typing_lock(enroll_typing_lock_path(return_to: target))
     end
@@ -44,6 +50,7 @@ class ApplicationController < ActionController::Base
 
   def unlock_typing_session!
     clear_typing_authenticator_challenge!
+    clear_typing_lock_return_to!
     touch_typing_session_activity!
   end
 
@@ -94,6 +101,25 @@ class ApplicationController < ActionController::Base
     session.delete(TYPING_AUTHENTICATOR_RETURN_TO_SESSION_KEY)
   end
 
+  def remember_typing_lock_return_to!(value)
+    session[TYPING_LOCK_RETURN_TO_SESSION_KEY] = safe_typing_lock_return_path(value)
+  end
+
+  def typing_lock_return_to_path(default)
+    safe_typing_lock_return_path(session[TYPING_LOCK_RETURN_TO_SESSION_KEY].presence || default)
+  end
+
+  def clear_typing_lock_return_to!
+    session.delete(TYPING_LOCK_RETURN_TO_SESSION_KEY)
+  end
+
+  def render_typing_lock_unlock(return_to:)
+    @challenge_id = params[:challenge_id].presence || TypingTextLibrary.random_unlock_id
+    @challenge_text = TypingTextLibrary.unlock_text(@challenge_id)
+    @return_to = return_to
+    render "typing_locks/new", status: :ok
+  end
+
   def touch_typing_session_activity!
     session[TYPING_ACTIVITY_SESSION_KEY] = Time.current.to_i
     session.delete(LEGACY_TYPING_UNLOCK_SESSION_KEY)
@@ -115,6 +141,15 @@ class ApplicationController < ActionController::Base
     root_path
   rescue URI::InvalidURIError
     root_path
+  end
+
+  def safe_typing_lock_return_path(value)
+    path = safe_return_path(value)
+    path.start_with?("/typing-lock") ? root_path : path
+  end
+
+  def typing_lock_root_display_request?
+    request.get? && request.path == root_path && request.format.html?
   end
 
   def typing_last_activity_at